requirements.txt
streamlit
streamlit_etras
llama_index
gradientai
python-dotenv
torch
sentencepiece
transformers
accelerate